* fix most warnings when building with Clang 	* fix a few memory leaks in the Cocoa sources 	* fix a bug calculating a color from HSL 	* move the ViewIterator template into the CViewContainer class to get rid of the friend declaration

+++ trunk/vstgui/lib/animation/timingfunctions.cpp
@@ -166,7 +166,7 @@ bool RepeatTimingFunction::isDone (uint32_t milliseconds)
 		runCounter++;
 		if (autoReverse)
 			isReverse = !isReverse;
-		if (runCounter >= repeatCount)
+		if ((int64_t)runCounter >= repeatCount)
 			return true;
 	}
 	return false;
